certpilot/requirements.txt
requests
whois
pillow
pytesseract
opencv-python
google-generativeai
playwright
beautifulsoup4
serpapi
numpy
